Update and rename pun.js to joke.js
